Add file bids_registration.py
This file allows to register two volumes together using a trained registration model given as input. It is quite similar to 3d_reg.py but more specifically designed to be used in a registration pipeline for datasets following BIDS convention. In the future 3d_reg.py and this file could be merged into one file, adding specific parser arguments.