Add option to force TF to use only 1 CPU
This option is set to True by default. Therefore, in the pipeline, the part performing the registration using a trained tensorflow model will use only one CPU per subject by default. If we want to use all the CPUs available for this part, the parser argument `--one-cpu-tf` should be set to `False`. These modifications are linked to issue #3620 in spinalcordtoolbox.