Add affine registration step when SC dice score below threshold
Implementation of an affine registration step prior to the model's registration when the model's registration didn't lead to a Dice score as high as wanted (and specified in the bash script). The addition of an affine registration to the pipeline is time consuming as this step needs the computation of the SC segmentations on the original volumes and is based on an optimization method. To avoid using the affine registration on any subjects, the argument MIN_SC_DICE_EXPECTED_PERC can be set to 0.