Split the original shell script and the one with the affine option into two files
Separate the two different scenarios (using an affine step when the results are not as high as expected or not) in order to let the user choose which option is preferred and to not add many useless code lines when the option without an affine step is preferred. This leads to the creation of a new shell script `pipeline_bids_register_evaluate_opt_affine.sh` that can be used when we want to take advantage of the affine registration under some conditions.