Compose the transformations obtained in the two steps
Change strategy of how the two steps process is done. Before it was done by calling twice the bids_registration.py in the script with two different models. Now, a new file has been created to replace the bids_registration.py in this case and in order to compute directly the two registration successively, allowing to access to the deformation field obtained after the first registration and compose it with the one obtained after the second registration, reducing the blurring effect of interpolation, when applying this composed deformation field to the moving volume.