Inference on input of any size or subvolumes
Implementation allowing to use the registration model on input of any size during inference without having to crop them and losing information before using the registration model. The user can choose whether to do the registration directly on the whole volume (more accurate but need more computational resources) or on subvolumes (with a size specified by the user). This can be chosen in the .py file but will be done through a config file later. When using subvolumes, the final deformation field to apply to the original volume is obtained by using all the deformation fields outputted by the model for the different subvolumes and doing a weighted average of the displacement vectors on the overlapping areas (giving more weights to voxels closer to the center of the subvolume than to the ones closer to the boundaries). The weighted average of the displacement vectors allow to obtain a smooth deformation field without sharp transitions delimiting the different subvolumes used. The registration using volume of any size during inference has been implemented for the normal registration strategy and for the two steps strategy.