Add infer any size or subvol for 3d_reg + correct preproc
Add the implementation allowing to use the registration model on input of any size during inference without having to crop them and losing information before using the registration model to the 3d_reg.py file.
Correction of how the volumes were preprocessed to set them to an isotropic resolution of 1mm.  Now use what is done in sct with sct_resample.