Use the compose warping field with whole vol
Use the compose warping field of the two registration steps to obtain the registered volume, in the case of the whole volume. This was done for the subvolumes but wasn't implemented when using the whole volume. The output warping field was composed but wasn't used to obtain the final registered volume.