Let the user choose interpolation for resampling
During the preprocessing step, the volumes need to be resampled to a 1 mm isotropic resolution. With the modifications done, we can choose whether we want a linear, spline or nearest neighbor interpolation to resample the volume. This is done with the config_inference file for the files linked to the pipeline and directly as parser argument for the 3d_reg.py file.