Change how warping field is transformed before being applied
When using the linear case, use directly the output of the model prediction. When using the nearest neighbor case, no upsampling of the warping field is performed anymore prior to using the transform model but the scaling factor of the transform model is specified. The output file representing the warping field is now therefore exactly the same as the one outputted by the model (without upsampling).