Generalize the pipeline
Generalize the framework for registration and evaluation to register any type of contrasts together.

The user needs to specify which modality (contrast) is used as the fixed image and which modality (contrast) will be registered to this fixed image (the moving image). In addition, the user should specify if data from multiple sessions are available.

Once this is done, the process will be fully automatic to register the moving image of the specified modality to the fixed image for each subject (and each session) of the dataset (organised following BIDS convention).