Fix how the warp field is returned
Fix how the warping field is returned to prepare for the integration in SCT.
It has been observed that the warping field that was returned in the original dimension of the input moving volume was not properly resampled since the preprocessing and postprocessing steps have been changed.
In addition, it has been observed that to be used with sct_apply_transfo(), the warping field needs to have five dimensions (additional time timension not present so far) with vector intent code.
Therefore, this PR aims to correct how the warping field obtained from the registration model is transformed before being saved (for now) or returned to be used in the pipeline (once integrated in SCT).