More general solution for vector orientation of warp field
This solution seems to work better than the previous one and should be general. It appears to work for both the spine generic dataset and the dataset of the Zurich project for sci segmentation. The voxel orientation of some of the modalities of the Zurich dataset were different than in the spine generic dataset (AIL vs RPI) allowing to highlight that the temporary solution was indeed working only for a subset of data (of RPI voxel orientation) whereas now the solution provided should normally work with any kind of data.