Add Jacobian (and det) computation on warping field
Compute the Jacobian (and its determinant) of the displacement vectors of the warping field.
This allows to get a better visualization of the warping field applied to the moving image and obtain some information on the type of deformation applied to each voxel. A Jacobian determinant higher than 1.0 represents voxel expansion, a Jacobian determinant between 0 and 1.0 represents voxel compression and a negative Jacobian determinant is representative of local folding.