Add choice to select registration evaluation tool
Add the option at the beginning of each script to choose what are the evaluation tools we are interested in. We can choose none, 1, 2, or all of the following evaluation options: metrics computed on the SC segmentation (Dice, Jaccard, ...); normalized mutual information; Jacobian determinant of the displacement vectors of the warping field.