Add transformation of moving volume with sct_apply_transfo
Transform the moving volume with sct_apply_transfo using the warping field outputted by the registration model. This allows to compare the results (registered volumes) obtained when using directly model.predict() or when using sct_apply_transfo.
For the moment, the registered volume obtained with model.predict() is the one used to compute the spinal cord segmentation and compute the metrics to assess the registration performance.