Separate QC files for SC seg and SC reg
Create one folder for the QC file that allows to observe if the segmentation of the SC went well and one folder to assess the SC registration.
In addition, remove a useless comparison between the moving image and the moved one as this can be done by going up and down in the QC report.